♻️ Refactor :  productdetailquery -> productoptionquery
