+++ src/blocks/main/(detail)/products/[productId]/info/BoardImagesSection.tsx
@@ -8,19 +8,19 @@ import { BundleBadge } from '@/domains/product/components/ProductCard/ProductIma
 import ProductImageSlide from '@/domains/product/components/ProductImageSlide';
 import ImageCounter from '@/domains/product/components/ProductImageSlide/ImgCounter';
 import useGetBoardDetailQuery from '@/domains/product/queries/useGetBoardDetailQuery';
-import useGetProductDetailQuery from '@/domains/product/queries/useGetProductDetailQuery';
+import useGetProductOptionQuery from '@/domains/product/queries/useGetProductOptionQuery';
 import PaddingWrapper from '@/shared/components/PaddingWrapper';
 
 const BoardImagesSection = ({ productId }: { productId: string }) => {
   const [swiperIndex, setSwiperIndex] = useState(0);
   const { data: boardDetail } = useGetBoardDetailQuery(productId);
-  const { data: ProductDetail } = useGetProductDetailQuery(productId);
+  const { data: productOption } = useGetProductOptionQuery(productId);
 
   return (
     <PaddingWrapper className="py-0">
       <div className="relative">
         <ProductImageSlide boardImages={boardDetail?.boardImages} onChange={setSwiperIndex} />
-        {ProductDetail?.boardIsBundled && (
+        {productOption?.boardIsBundled && (
           <div className="absolute top-[10px] left-[10px] z-10 ">
             <BundleBadge />
           </div>
